storage/nvme: Fix different size operands in bitwise operations
NVME_ADMIN_CRIOCQ_QID and NVME_ADMIN_CRIOCQ_QSIZE macros are used
for the cdw10 field, which is of type uint32_t. As a result,
the following code

 sq->cdw10 |= NVME_ADMIN_CRIOCQ_QID(qid);

for uint16_t qid leads to an error in Klocwork static analysis.
Fix the type returned by these macros (as well as NVME_ADMIN_CRIOCQ_*).

BRANCH=none
BUG=none
TEST=Boot to OS on Nivviks

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/4005406
Commit-Queue: Tarun Tuli <taruntuli@google.com>
Tested-by: Tarun Tuli <taruntuli@google.com>

+++ src/drivers/storage/nvme.h
@@ -177,16 +177,16 @@ typedef uint32_t NVME_CQHDBL;
 
 /* NVMe Admin Cmd Opcodes */
 #define NVME_ADMIN_CRIOSQ_OPC	1
-#define NVME_ADMIN_CRIOSQ_QID(x)	(x)
-#define NVME_ADMIN_CRIOSQ_QSIZE(x)	(((x)-1) << 16)
-#define NVME_ADMIN_CRIOSQ_CQID(x)	((x) << 16)
+#define NVME_ADMIN_CRIOSQ_QID(x)	((uint32_t)x)
+#define NVME_ADMIN_CRIOSQ_QSIZE(x)	(((uint32_t)(x)-1) << 16)
+#define NVME_ADMIN_CRIOSQ_CQID(x)	(((uint32_t)x) << 16)
 
 #define NVME_ADMIN_SETFEATURES_OPC	9
 #define NVME_ADMIN_SETFEATURES_NUMQUEUES	7
 
 #define NVME_ADMIN_CRIOCQ_OPC	5
-#define NVME_ADMIN_CRIOCQ_QID(x)	(x)
-#define NVME_ADMIN_CRIOCQ_QSIZE(x)	(((x)-1) << 16)
+#define NVME_ADMIN_CRIOCQ_QID(x)	((uint32_t)x)
+#define NVME_ADMIN_CRIOCQ_QSIZE(x)	(((uint32_t)(x)-1) << 16)
 
 #define NVME_ADMIN_GET_LOG_PAGE	0x02
 #define NVME_ADMIN_IDENTIFY_OPC	0x06
